Copied from ldlite 2.3.

+++ OptionsDialog.cpp
@@ -218,9 +218,9 @@ void COptionsDialog::OnDeltaposScaleSmalldelta(NMHDR* pNMHDR, LRESULT* pResult)
 	NM_UPDOWN* pNMUpDown = (NM_UPDOWN*)pNMHDR;
 	// TODO: Add your control notification handler code here
 	UpdateData(TRUE);
-	m_scale += -(pNMUpDown->iDelta) * 0.1;
+	m_scale += -(pNMUpDown->iDelta) * (float)0.1;
 	if( m_scale < 0.1 ) {
-		m_scale = 0.1;
+		m_scale = (float)0.1;
 	}
 
 	OnChanged();
@@ -237,7 +237,7 @@ void COptionsDialog::OnDeltaposScaleBigdelta(NMHDR* pNMHDR, LRESULT* pResult)
 	UpdateData(TRUE);
 	m_scale += -(pNMUpDown->iDelta) * 1;
 	if( m_scale < 0.1 ) {
-		m_scale = 0.1;
+		m_scale = (float)0.1;
 	}
 
 	OnChanged();
